# project-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Provider Object Storage API
  version: 0.1.0
  description: |
    Initial contract for the implemented control-plane and custom storage API.
    This draft covers the organization membership and role administration
    vertical slice together with provider setup and storage clients.
    Organization management routes require a user JWT and the
    x-organization-id header. Platform routes require a global role with the
    platform permission and do not accept an organization header as scope.
    Storage routes require an API credential in the form keyId.secret and
    enforce namespace and folder-grant boundaries.
servers:
  - url: http://localhost:4000
    description: Local API
...
paths:
...
  /v1/organizations/{organizationId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: getOrganization
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Organization summary.
          content:
            application/json:
              schema:
                type: object
                required: [organization]
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        '200':
          description: Updated organization.
          content:
            application/json:
              schema:
                type: object
                required: [organization]
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/organizations/{organizationId}/activity-logs:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: listActivityLogs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
            pattern: '^\d{1,20}$'
        - name: limit
          in: query
          schema:
            type: string
            pattern: '^\d{1,3}$'
            default: '50'
        - name: action
          in: query
          schema:
            type: string
            maxLength: 120
        - name: entityType
          in: query
          schema:
            type: string
            maxLength: 120
      responses:
        '200':
          description: Activity log page.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLog'
                  nextCursor:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/organizations/{organizationId}/usage:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: getOrganizationUsage
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: string, pattern: '^\d{1,3}$', default: '50' }
      responses:
        '200':
          description: Organization usage totals and a bounded namespace page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsagePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/organizations/{organizationId}/usage/export:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: exportOrganizationUsage
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: string, pattern: '^\d{1,3}$', default: '50' }
      responses:
        '200':
          description: Bounded CSV usage export. Continue with X-Next-Cursor when present.
          headers:
            X-Next-Cursor:
              schema: { type: string, format: uuid }
          content:
            text/csv:
              schema: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/organizations/{organizationId}/members:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: listOrganizationMembers
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Organization members and assigned role names.
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Organizations]
      operationId: addOrganizationMember
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberAddRequest'
      responses:
        '201':
          description: Member added.
          content:
            application/json:
              schema:
                type: object
                required: [member]
                properties:
                  member: { $ref: '#/components/schemas/Member' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/organizations/{organizationId}/members/{membershipId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/MembershipId'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    patch:
      tags: [Organizations]
      operationId: updateMemberStatus
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberStatusUpdateRequest'
      responses:
        '200':
          description: Member status updated.
          content:
            application/json:
              schema:
                type: object
                required: [member]
                properties:
                  member: { $ref: '#/components/schemas/Member' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/organizations/{organizationId}/members/{membershipId}/roles:
    post:
      tags: [Organizations]
      operationId: assignMemberRole
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationIdPath'
        - $ref: '#/components/parameters/MembershipId'
        - $ref: '#/components/parameters/OrganizationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignmentRequest'
      responses:
        '200':
          description: Updated member role assignment.
          content:
            application/json:
              schema:
                type: object
                required: [member]
                properties:
                  member: { $ref: '#/components/schemas/Member' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/organizations/{organizationId}/roles:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    get:
      tags: [Organizations]
      operationId: listOrganizationRoles
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Organization roles and permission keys.
          content:
            application/json:
              schema:
                type: object
                required: [roles]
                properties:
                  roles:
                    type: array
                    items: { $ref: '#/components/schemas/Role' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Organizations]
      operationId: createOrganizationRole
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Role created.
          content:
            application/json:
              schema:
                type: object
                required: [role]
                properties:
                  role: { $ref: '#/components/schemas/Role' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/organizations/{organizationId}/roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationIdPath'
      - $ref: '#/components/parameters/RoleId'
      - $ref: '#/components/parameters/OrganizationIdHeader'
    patch:
      tags: [Organizations]
      operationId: updateOrganizationRole
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role updated.
          content:
            application/json:
              schema:
                type: object
                required: [role]
                properties:
                  role: { $ref: '#/components/schemas/Role' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Organizations]
      operationId: deleteOrganizationRole
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Role deleted.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived dashboard access token.
    storageCredential:
      type: http
      scheme: bearer
      bearerFormat: keyId.secret
      description: API credential scoped to one organization and namespace.
    metricsBearer:
      type: http
      scheme: bearer
      description: Deployment-provided metrics scrape token.
  parameters:
    OrganizationIdHeader:
      name: x-organization-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    OrganizationIdPath:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Missing, invalid, expired, revoked, or out-of-scope credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource was not found in the caller's tenant scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource state or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
...
    Organization:
      type: object
      required: [id, name, slug, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [ACTIVE, SUSPENDED, CLOSED] }
    Member:
      type: object
      required: [id, userId, email, status, roles]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        status: { type: string, enum: [INVITED, ACTIVE, DISABLED] }
        roles:
          type: array
          items: { type: string }
    Role:
      type: object
      required: [id, name, isSystem, permissionKeys]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        isSystem: { type: boolean }
        permissionKeys:
          type: array
          uniqueItems: true
          items: { type: string }
    MemberAddRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        roleId: { type: string, format: uuid }
    MemberStatusUpdateRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { type: string, enum: [ACTIVE, DISABLED] }
    RoleAssignmentRequest:
      type: object
      required: [roleId]
      additionalProperties: false
      properties:
        roleId: { type: string, format: uuid }
    RoleCreateRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        permissionKeys:
          type: array
          uniqueItems: true
          items: { type: string }
          default: []
    RoleUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        permissionKeys:
          type: array
          uniqueItems: true
          items: { type: string }
    ActivityLog:
      type: object
      required: [id, action, entityType, requestId, createdAt, metadata, actor]
      properties:
        id: { type: string }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string, format: uuid, nullable: true }
        requestId: { type: string }
        createdAt: { type: string, format: date-time }
        metadata: {}
        actor:
          type: object
          nullable: true
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
    UsagePageResponse:
      type: object
      required: [usage]
      properties:
        usage:
          type: object
          required: [namespaces, nextCursor, totals]
          properties:
            nextCursor: { type: string, format: uuid, nullable: true }
            totals: { $ref: '#/components/schemas/UsageTotals' }
            namespaces:
              type: array
              items: { $ref: '#/components/schemas/NamespaceUsage' }
    UsageTotals:
      type: object
      required: [usedBytes, reservedBytes, objectCount, activeUploadCount, requestCount, ingressBytes, egressBytes]
      properties:
        usedBytes: { type: string, pattern: '^\d+$' }
        reservedBytes: { type: string, pattern: '^\d+$' }
        objectCount: { type: string, pattern: '^\d+$' }
        activeUploadCount: { type: integer, minimum: 0 }
        requestCount: { type: string, pattern: '^\d+$' }
        ingressBytes: { type: string, pattern: '^\d+$' }
        egressBytes: { type: string, pattern: '^\d+$' }
    NamespaceUsage:
      allOf:
        - $ref: '#/components/schemas/UsageTotals'
        - type: object
          required: [namespaceId, name, slug, status, quotaBytes, updatedAt]
          properties:
            namespaceId: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
            status: { type: string }
            quotaBytes: { type: string, pattern: '^\d+$' }
            updatedAt: { type: string, format: date-time, nullable: true }
...